Java
This is a GUI application for the APPLE STORE, which is for the cashier, manager, and the products and stock management system. In this application, the Manager can create new accounts for the new employees and Manager can add products, and stocks for the store. Only the Manager can add products and he or she can add, update, delete, search, and view the products and stock. The cashier can only view and search the products and stock.